Render progressions as tabs
